sgetline updates
git-svn-id: file:///source/coreland/corelib/trunk@667 6e7d7e42-f3d6-da11-80d8-00d0b73f4485

+++ sgetline.c
@@ -2,9 +2,9 @@
 
 void sgetline_init(sgetline *s,
                    char *sbuf,
-                   unsigned int slen,
+                   unsigned long slen,
                    char *bbuf,
-                   unsigned int blen,
+                   unsigned long blen,
                    long (*op)(int, char *, unsigned long),
                    int fd)
 {
@@ -19,7 +19,7 @@ void sgetline_init(sgetline *s,
   s->b.pos = 0;
 }
 
-void sgetline_initbuf(sgetline *s, char *sbuf, unsigned int slen, buffer *b)
+void sgetline_initbuf(sgetline *s, char *sbuf, unsigned long slen, buffer *b)
 {
   s->s = sbuf;
   s->len = 0;
